refactor: extract theme tokens, split Contact, de-duplicate the Skills filter
Three extractions, all behaviour-preserving - the prerender still injects an
identical 86919 chars into #root, which is the strongest evidence available
that the emitted class strings did not move.

shared/theme/tokens.ts (was utils/cardStyles.ts) now owns every themed class
pair: headingText, bodyText, accentText, filterChip, inputField, primaryButton,
disabledButton alongside the existing cardSurface. 21 inline darkMode ternaries
across thirteen files became calls. tokens.test.ts asserts each helper returns
a DIFFERENT string per theme - a helper that ignored its argument would
otherwise render the wrong theme everywhere with no test failing.

Left inline on purpose, because they are decisions and not drift: the Hero's
heavier surface, the red ErrorBoundary, the green contact-success box, and the
Footer's brand-icon colours.

sections/skills/SkillFilterButton.tsx collapses the 'All' tab and the seven
category tabs, which were two JSX blocks carrying the same ten-line class
expression verbatim.

sections/contact/ splits 408 lines into useContactForm.ts (state, validation,
captcha lifecycle, submit), ContactField.tsx and a presentation-only
Contact.tsx (240 lines). reCAPTCHA behaviour is untouched: clearCaptcha still
pairs the widget reset with the state clear, and the widgetRef still resets the
spent single-use token.
