+++ src/layout/Footer.tsx
@@ -1,14 +1,13 @@
 import { useTheme } from '@/shared/hooks/useTheme';
 import { FaReact } from 'react-icons/fa';
 import { SiFramer, SiTailwindcss, SiTypescript } from 'react-icons/si';
+import { headingText } from '@/shared/theme/tokens';
 
 const Footer = () => {
   const { darkMode } = useTheme();
 
   return (
-    <footer
-      className={`p-4 text-center ${darkMode ? 'text-white' : 'text-gray-900'}`}
-    >
+    <footer className={`p-4 text-center ${headingText(darkMode)}`}>
       <p className='text-xs sm:text-base'>
         © Aleksandar Trenchevski {new Date().getFullYear()}. All rights
         reserved.
