+++ src/sections/education/Education.tsx
@@ -3,16 +3,21 @@ import { motion } from 'motion/react';
 import education from '@/data/education.json';
 import publications from '@/data/publications.json';
 import { staggerContainer, slideUp } from '@/shared/utils/animationVariants';
-import { cardSurface } from '@/shared/theme/tokens';
+import {
+  accentText,
+  bodyText,
+  cardSurface,
+  headingText,
+} from '@/shared/theme/tokens';
 
 const Education = () => {
   const { darkMode } = useTheme();
-  const accent = darkMode ? 'text-cyan-400' : 'text-fuchsia-600';
+  const accent = accentText(darkMode);
   const card = `rounded-lg shadow-lg p-5 sm:p-6 ${cardSurface(darkMode)}`;
 
   return (
     <motion.section
-      className={`flex justify-center animate-smoothFadeIn ${darkMode ? 'text-white' : 'text-gray-700'}`}
+      className={`flex justify-center animate-smoothFadeIn ${bodyText(darkMode)}`}
       initial='hidden'
       whileInView='visible'
       viewport={{ once: true }}
@@ -20,7 +25,7 @@ const Education = () => {
     >
       <div className='w-full max-w-lg sm:max-w-3xl'>
         <motion.h2
-          className={`text-2xl sm:text-3xl lg:text-4xl font-bold text-center mb-8 sm:mb-12 ${darkMode ? 'text-white' : 'text-gray-900'}`}
+          className={`text-2xl sm:text-3xl lg:text-4xl font-bold text-center mb-8 sm:mb-12 ${headingText(darkMode)}`}
           variants={slideUp}
         >
           Education
@@ -51,7 +56,7 @@ const Education = () => {
         {publications.length > 0 && (
           <motion.div className='mt-10' variants={slideUp}>
             <h3
-              className={`text-xl sm:text-2xl font-bold text-center mb-6 ${darkMode ? 'text-white' : 'text-gray-900'}`}
+              className={`text-xl sm:text-2xl font-bold text-center mb-6 ${headingText(darkMode)}`}
             >
               Publications
             </h3>
